Added missing docs
