+++ KMART/Model/Configuration.swift
@@ -5,39 +5,46 @@
 //  Created by Nindi Gill on 14/2/21.
 //
 
-import ArgumentParser
 import Foundation
 import Yams
 
-enum ConfigurationType: String, ExpressibleByArgument {
-    case json = "JSON"
-    case plist = "Property List"
-    case yaml = "YAML"
-}
-
-enum ConfigurationError: Error {
-    case invalidFile
-}
-
+/// Struct used to hold all configuration options.
 struct Configuration {
+    /// Report name.
     var name: String = "\(String.appName.capitalized) Report"
+    /// Jamf Pro URL.
     var url: String = ""
+    /// Jamf Pro API credentials.
     var credentials: String = ""
+    /// Jamf Pro API basic authorization token.
     var authorization: String {
         "Basic \(credentials)"
     }
+    /// Number of concurrent API requests.
     var requests: Int = 4
+    /// Number of seconds before a timeout.
     var timeout: Double = 10
+    /// List of report types to report on.
     var reports: [ReportType] = []
+    /// Additional report options.
     var reportOptions: [ReportOptionType: Int] = [:]
+    /// List of output types to export.
     var output: [OutputType: String] = [:]
+    /// Email configuration object.
     var email: EmailConfiguration = EmailConfiguration([:])
+    /// Slack configuration object.
     var slack: SlackConfiguration = SlackConfiguration([:])
+    /// List of endpoints to report on, derived from the requested report types.
     var endpoints: [Endpoint] {
         let endpoints: [Endpoint] = reports.flatMap { $0.endpoints }
         return Array(Set(endpoints)).sorted { $0.identifier < $1.identifier }
     }
 
+    /// Initialize a Configuration struct by passing in a custom `ConfigurationType` and path to a configuration file.
+    ///
+    /// - Parameters:
+    ///   - type: Custom `ConfigurationType` (ie. `.json`, `.plist`, `.yaml`).
+    ///   - path: Path to a configuration file.
     init?(_ type: ConfigurationType, path: String) {
 
         let url: URL = URL(fileURLWithPath: path)
@@ -78,6 +85,10 @@ struct Configuration {
         configureSlack(dictionary["slack"] as? [String: Any] ?? [: ])
     }
 
+    /// Set the default Jamf API configuration.
+    ///
+    /// - Parameters:
+    ///   - dictionary: The dictionary containing the **Report** configuration options.
     private mutating func configureSetup(_ dictionary: [String: Any]) -> Bool {
 
         if let string: String = dictionary["name"] as? String {
@@ -109,6 +120,10 @@ struct Configuration {
         return true
     }
 
+    /// Set the **Reports** configuration.
+    ///
+    /// - Parameters:
+    ///   - dictionary: The dictionary containing the **Report** configuration options.
     private mutating func configureReports(_ dictionary: [String: Bool]) {
 
         for key in dictionary.keys {
@@ -125,6 +140,10 @@ struct Configuration {
         reports.sort { $0.identifier < $1.identifier }
     }
 
+    /// Set the **Report Options** configuration.
+    ///
+    /// - Parameters:
+    ///   - dictionary: The dictionary containing the **Report** configuration options.
     private mutating func configureReportOptions(_ dictionary: [String: Int]) {
 
         for key in dictionary.keys.sorted() {
@@ -139,6 +158,10 @@ struct Configuration {
         }
     }
 
+    /// Set the **Output** configuration.
+    ///
+    /// - Parameters:
+    ///   - dictionary: The dictionary containing the **Output** configuration options.
     private mutating func configureOutput(_ dictionary: [String: String]) {
 
         for key in dictionary.keys.sorted() {
@@ -152,10 +175,18 @@ struct Configuration {
         }
     }
 
+    /// Set the **Email** configuration.
+    ///
+    /// - Parameters:
+    ///   - dictionary: The dictionary containing the **Email** configuration options.
     private mutating func configureEmail(_ dictionary: [String: Any]) {
         email = EmailConfiguration(dictionary)
     }
 
+    /// Set the **Slack** configuration.
+    ///
+    /// - Parameters:
+    ///   - dictionary: The dictionary containing the **Slack** configuration options.
     private mutating func configureSlack(_ dictionary: [String: Any]) {
         slack = SlackConfiguration(dictionary)
     }
