+++ KMART/Model/Enums/ConfigurationError.swift
@@ -1,0 +1,12 @@
+//
+//  ConfigurationError.swift
+//  KMART
+//
+//  Created by Nindi Gill on 13/12/21.
+//
+
+import Foundation
+
+enum ConfigurationError: Error {
+    case invalidFile
+}
