+++ KMART/Model/Enums/Endpoint.swift
