+++ KMART/Model/Enums/OutputType.swift
