+++ KMART/Model/Enums/ReportOptionType.swift
