+++ KMART/Model/Enums/ReportType.swift
