+++ KMART/Model/Enums/ScriptType.swift
